Change directory back after Gulp Cordova tasks
We need to change the current working directory to use Cordova Lib.
Once we have finished we need to ensure we change the curent working directory
back, even if we error. We also need to make parts of the Cordova tasks
synchronous so that other tasks don't run with the wrong working directory.